refactor(rdw): move schema validation to codec
